fix: 코드 리뷰 반영 #16

+++ src/main/kotlin/com/dh/baro/look/domain/repository/LookProductRepository.kt
@@ -3,6 +3,6 @@ package com.dh.baro.look.domain.repository
 import com.dh.baro.look.domain.LookProduct
 import org.springframework.data.jpa.repository.JpaRepository
 
-interface LookProductRepository : JpaRepository<LookProduct, Int> {
+interface LookProductRepository : JpaRepository<LookProduct, Long> {
     fun findByLookIdOrderByDisplayOrderAsc(lookId: Long): List<LookProduct>
 }
